+++ spicerack/redfish.py
@@ -995,11 +995,9 @@ class RedfishDell(Redfish):
             spicerack.redfish.RedfishTaskNotCompletedError: if unable to fetch the dumped results.
 
         """
-        data: dict = {"ExportFormat": "JSON", "ShareParameters": {"Target": target.value}}
-        if self.generation >= self.idrac_10_min_gen:
-            # Hardcoded for now because the other possible values (NFS, CIFS, HTTP, HTTPS) requires additional settings
-            data["ShareParameters"]["ShareType"] = "LOCAL"
-
+        # iDRAC 10 wants Target as a list
+        target_value = [target.value] if self.generation >= self.idrac_10_min_gen else target.value
+        data: dict = {"ExportFormat": "JSON", "ShareParameters": {"Target": target_value}}
         task_uri = self.submit_task(f"{self.scp_base_uri}.ExportSystemConfiguration", data)
         # Wait before starting to poll for the task, so that a quick task can complete before the first attempt.
         time.sleep(5)
@@ -1040,16 +1038,14 @@ class RedfishDell(Redfish):
         else:
             uri = "ImportSystemConfiguration"
 
+        # iDRAC 10 wants Target as a list
+        target_value = [scp.target.value] if self.generation >= self.idrac_10_min_gen else scp.target.value
         data: dict = {
             "ImportBuffer": json.dumps(scp.config),  # The API requires a JSON-encoded string inside a JSON payload.
-            "ShareParameters": {"Target": scp.target.value},
+            "ShareParameters": {"Target": target_value},
             "HostPowerState": power_state.value,
             "ShutdownType": reboot.value,
         }
-        if self.generation >= self.idrac_10_min_gen:
-            # Hardcoded for now because the other possible values (NFS, CIFS, HTTP, HTTPS) requires additional settings
-            data["ShareParameters"]["ShareType"] = "LOCAL"
-
         task_id = self.submit_task(f"{self.scp_base_uri}.{uri}", data)
 
         return self.poll_task(task_id)